+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Doubled the speed of encoding bytes to base64
+
 ## [0.2.0] - 2019-08-12
 ### Added
 - Add a new `Presence` class to support all presence properties